requirements.txt
torch
torchvision
transformers
tqdm
spacy
pandas
numpy
Pillow
clip-benchmark
datasets==3.6.0
huggingface-hub==0.34.4